refactor(vendo): wire route table — misc + single-table dispatch (kill-list B4)
